+++ src/main/features/claim-documents/index.ts
@@ -4,7 +4,20 @@ import * as path from 'path'
 import { Paths } from 'claim-documents/paths'
 
 import { RouterFinder } from 'shared/router/routerFinder'
-import { claimIssueRequestHandler } from 'claim/index'
+import { OAuthHelper } from 'idam/oAuthHelper'
+import { AuthorizationMiddleware } from 'idam/authorizationMiddleware'
+
+export function claimIssueRequestHandler (): express.RequestHandler {
+  function accessDeniedCallback (req: express.Request, res: express.Response): void {
+    res.redirect(OAuthHelper.forLogin(req, res))
+  }
+
+  const requiredRoles = [
+    'citizen'
+  ]
+  const unprotectedPaths = []
+  return AuthorizationMiddleware.requestHandler(requiredRoles, accessDeniedCallback, unprotectedPaths)
+}
 
 export class Feature {
   enableFor (app: express.Express) {
