+++ src/main/features/claim/index.ts
@@ -3,33 +3,20 @@ import * as path from 'path'
 
 import { Paths } from 'claim/paths'
 
-import { AuthorizationMiddleware } from 'idam/authorizationMiddleware'
 import { ClaimEligibilityGuard } from 'claim/guards/claimEligibilityGuard'
 import { RouterFinder } from 'shared/router/routerFinder'
 import { DraftMiddleware } from '@hmcts/cmc-draft-store-middleware'
 import { DraftService } from 'services/draftService'
 import { DraftClaim } from 'drafts/models/draftClaim'
-import { OAuthHelper } from 'idam/oAuthHelper'
-
-export function claimIssueRequestHandler (): express.RequestHandler {
-  function accessDeniedCallback (req: express.Request, res: express.Response): void {
-    res.redirect(OAuthHelper.forLogin(req, res))
-  }
-
-  const requiredRoles = [
-    'citizen'
-  ]
-  const unprotectedPaths = []
-  return AuthorizationMiddleware.requestHandler(requiredRoles, accessDeniedCallback, unprotectedPaths)
-}
+// import { claimIssueRequestHandler } from 'claim-documents/index'
 
 export class Feature {
   enableFor (app: express.Express) {
     if (app.settings.nunjucksEnv && app.settings.nunjucksEnv.globals) {
       app.settings.nunjucksEnv.globals.ClaimPaths = Paths
     }
-
-    app.all('/claim/*', claimIssueRequestHandler())
+    // commented out as claim document index.ts will handle this
+    // app.all('/claim/*', claimIssueRequestHandler())
     app.all(/^\/claim\/(?!start|amount-exceeded|new-features-consent|.+\/confirmation|.+\/receipt|.+\/draftReceipt|.+\/sealed-claim).*$/,
       DraftMiddleware.requestHandler(new DraftService(), 'claim', 100, (value: any): DraftClaim => {
         return new DraftClaim().deserialize(value)
